Revert "Move tests"
This reverts commit 96a00778b6da336cdda357bd729645cfa42d9b0f.
