+++ test/exclude_time_period_test.js
@@ -1,10 +1,10 @@
 import React from 'react'
-import * as utils from '../../src/date_utils'
+import * as utils from '../src/date_utils'
 // import ReactDOM from 'react-dom'
 // import TestUtils from 'react-addons-test-utils'
 import { mount } from 'enzyme'
-import DatePicker from '../../src/datepicker.jsx'
-// import Time from '../../src/time.jsx'
+import DatePicker from '../src/datepicker.jsx'
+// import Time from '../src/time.jsx'
 
 describe('DatePicker', () => {
   let sandbox
